feat: made sure CSR is consistent
