+++ app/layout.tsx
@@ -1,6 +1,8 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
+import Navbar from "@/components/Navbar";
+import Footer from "@/components/Footer";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -16,7 +18,11 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className='w-full h-screen text-center'>{children}</body>
+      <body className='w-full h-screen text-center'>
+        <Navbar/>
+        {children}
+        <Footer/>
+      </body>
     </html>
   );
 }
